light mode is the default, add mobilenetv2ssdlite example
