+++ examples/CMakeLists.txt
@@ -16,4 +16,11 @@ target_link_libraries(fasterrcnn ncnn ${OpenCV_LIBS})
 add_executable(yolov2 yolov2.cpp)
 target_link_libraries(yolov2 ncnn ${OpenCV_LIBS})
 
-add_subdirectory(ssd)
+add_executable(mobilenetv2ssdlite mobilenetv2ssdlite.cpp)
+target_link_libraries(mobilenetv2ssdlite ncnn ${OpenCV_LIBS})
+
+add_executable(mobilenetssd mobilenetssd.cpp)
+target_link_libraries(mobilenetssd ncnn ${OpenCV_LIBS})
+
+add_executable(squeezenetssd squeezenetssd.cpp)
+target_link_libraries(squeezenetssd ncnn ${OpenCV_LIBS})
