+++ examples/squeezenet.cpp
@@ -32,7 +32,6 @@ static int detect_squeezenet(const cv::Mat& bgr, std::vector<float>& cls_scores)
     in.substract_mean_normalize(mean_vals, 0);
 
     ncnn::Extractor ex = squeezenet.create_extractor();
-    ex.set_light_mode(true);
 
     ex.input("data", in);
 
@@ -75,6 +74,12 @@ static int print_topk(const std::vector<float>& cls_scores, int topk)
 
 int main(int argc, char** argv)
 {
+    if (argc != 2)
+    {
+        fprintf(stderr, "Usage: %s [imagepath]\n", argv[0]);
+        return -1;
+    }
+
     const char* imagepath = argv[1];
 
     cv::Mat m = cv::imread(imagepath, CV_LOAD_IMAGE_COLOR);
@@ -91,4 +96,3 @@ int main(int argc, char** argv)
 
     return 0;
 }
-
